requirements.txt
aiofiles==23.1.0
anyio==3.6.2
async-timeout==4.0.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
fastapi==0.92.0
filelock==3.9.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
locust==2.14.2
loguru==0.6.0
MarkupSafe==2.1.2
msgpack==1.0.4
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
Pillow==9.4.0
pkg_resources==0.0.0
psutil==5.9.4
pydantic==1.10.5
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
roundrobin==0.0.4
scikit-learn==1.2.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
